refactor(editor): move CodeEditor component to editor directory
